MLops_Multiclass_Rakuten/rakuten_api/app/requirements.txt
# --- Framework & Serveur ---
fastapi==0.79.0
uvicorn==0.18.2
starlette==0.19.1
python-multipart==0.0.5

# --- Machine Learning (Le cœur) ---
tensorflow==2.9.1
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
nltk==3.7

# --- MLflow & S3 (Indispensable pour vos artefacts) ---
mlflow==1.28.0
boto3==1.24.0

# --- Base de données (Mongo/Motor) ---
pymongo==4.2.0
motor==3.0.0

# --- Sécurité & Utilitaires ---
pydantic==1.9.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==1.2.1
requests==2.28.1
rsa==4.7.2
tabulate==0.8.10

# --- Compatibilité (On laisse pip gérer les versions minimales) ---
typing-extensions>=4.1.1
six>=1.16.0